Bugfixes, Cleanup
