--- clone_dependencies.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-git clone https://github.com/J4n1X/hope.git include/hope
-git clone https://github.com/OneLoneCoder/olcPixelGameEngine include/olcPixelGameEngine
-git clone https://github.com/btzy/nativefiledialog-extended include/nativefiledialog-extended
