+++ src/libElecSim/GridTile.h
@@ -7,108 +7,9 @@
 #include <vector>
 
 #include "olcPixelGameEngine.h"
+#include "Common.h"
 
 namespace ElecSim {
-
-constexpr int GRIDTILE_BYTESIZE =
-    sizeof(int) * 4;  // TileId + Facing + PosX + PosY
-
-enum class Direction { Top = 0, Right = 1, Bottom = 2, Left = 3, Count = 4 };
-constexpr std::string_view DirectionToString(Direction dir) {
-  switch (dir) {
-    case Direction::Top:
-      return "Top";
-    case Direction::Right:
-      return "Right";
-    case Direction::Bottom:
-      return "Bottom";
-    case Direction::Left:
-      return "Left";
-    default:
-      return "Unknown";
-  }
-}
-static constexpr std::array<Direction, static_cast<int>(Direction::Count)>
-    AllDirections = {Direction::Top, Direction::Right, Direction::Bottom,
-                     Direction::Left};
-
-// This is used only to inform if an update happened or not
-struct TileSideStates
-    : private std::array<bool, static_cast<int>(Direction::Count)> {
- private:
-  typedef bool T;
-  typedef std::array<bool, static_cast<int>(Direction::Count)> array;
- public:
-  using array::fill;
-  using array::operator[];
-  using array::begin;
-  using array::end;
-  using array::size;
-  TileSideStates() { this->fill(false); }
-  TileSideStates(std::initializer_list<std::tuple<Direction, bool>> dirs) {
-    this->fill(false);
-    for (const auto& [dir, val] : dirs) {
-      if (dir >= Direction::Top && dir < Direction::Count) {
-        (*this)[static_cast<int>(dir)] = val;
-      }
-    }
-  }
-  bool& operator[](Direction dir) { return (*this)[static_cast<int>(dir)]; }
-  const bool& operator[](Direction dir) const {
-    return std::array<bool, static_cast<int>(Direction::Count)>::operator[](static_cast<int>(dir));
-  }
-};
-
-// Direction utility functions
-inline Direction FlipDirection(Direction dir) {
-  switch (dir) {
-    case Direction::Top:
-      return Direction::Bottom;
-    case Direction::Bottom:
-      return Direction::Top;
-    case Direction::Left:
-      return Direction::Right;
-    case Direction::Right:
-      return Direction::Left;
-    default:
-      return dir;
-  }
-}
-
-struct SignalEvent {
-  olc::vi2d sourcePos;
-  Direction fromDirection;
-  bool isActive;
-
-  SignalEvent(olc::vi2d pos, Direction toDirection, bool active)
-      : sourcePos(pos),
-        fromDirection(FlipDirection(toDirection)),
-        isActive(active){}
-  
-  // Add a copy constructor to ensure the visitedPositions gets copied
-  SignalEvent(const SignalEvent& other)
-      : sourcePos(other.sourcePos),
-        fromDirection(other.fromDirection),
-        isActive(other.isActive) {}
-};
-
-// Forward declare GridTile for UpdateEvent
-class GridTile;
-
-// Event for the update queue with priority ordering
-struct UpdateEvent {
-  std::shared_ptr<GridTile> tile;
-  SignalEvent event;
-  uint32_t updateCycleId;
-
-  UpdateEvent(std::shared_ptr<GridTile> t, const SignalEvent& e, int id)
-      : tile(t), event(e), updateCycleId(id) {}
-
-  bool operator<(const UpdateEvent& other) const {
-    return updateCycleId < other.updateCycleId;
-  }
-};
-
 // Base tile class
 // TODO: I want to add a Construct() function that creates a new tile of any
 // type, maybe with templates? A factory?
